% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootei-package.R
\docType{package}
\name{bootei-package}
\alias{bootei-package}
\title{bootei: BOOTEI permutation tests}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
